requirements.txt
# AI-ARTWORKS QUANTUM ULTRA - Maximum Requirements
# Ultimate performance with all cutting-edge libraries

# Quantum Computing
qiskit>=1.0.0
qiskit-aer>=0.14.0
qiskit-machine-learning>=0.7.0
cirq>=1.3.0
pennylane>=0.35.0
strawberryfields>=0.23.0
amazon-braket-sdk>=1.80.0

# GUI - Maximum QML and 3D
PySide6>=6.7.0
PySide6-Addons>=6.7.0
PySide6-Essentials>=6.7.0
PyQt6>=6.7.0
PyQt6-3D>=6.7.0
PyQt6-Charts>=6.7.0
PyQt6-DataVisualization>=6.7.0
PyQt6-NetworkAuth>=6.7.0
PyQt6-WebEngine>=6.7.0

# Advanced Graphics
moderngl>=5.10.0
vispy>=0.14.0
pyglet>=2.0.0
pycairo>=1.25.0
pyopengl>=3.1.7
vtk>=9.3.0
mayavi>=4.8.0
pyvista>=0.43.0

# Neural Networks - Maximum
torch>=2.2.0+cu121
torchvision>=0.17.0+cu121
torchaudio>=2.2.0+cu121
tensorflow>=2.16.0
jax[cuda12_pip]>=0.4.25
flax>=0.8.0
optax>=0.2.0

# Transformers - Latest Models
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
auto-gptq>=0.7.0
optimum>=1.18.0
peft>=0.10.0

# Advanced AI/ML
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
llama-index>=0.10.0
guidance>=0.1.0
semantic-kernel>=1.0.0
autogen>=0.2.0

# Vector Databases - Maximum
chromadb>=0.5.0
faiss-gpu>=1.7.4
qdrant-client>=1.9.0
weaviate-client>=4.5.0
pinecone-client>=3.2.0
milvus>=2.4.0
lance>=0.10.0

# Image Processing - Maximum
Pillow>=10.3.0
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
scikit-image>=0.23.0
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
kornia>=0.7.0
albumentations>=1.4.0

# Audio Processing - Maximum
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
pyaudio>=0.2.14
sounddevice>=0.4.6
spleeter>=2.4.0
openai-whisper>=20231117
bark>=1.0.0
audiocraft>=1.3.0

# Scientific Computing - Maximum
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
polars>=0.20.0
numba>=0.59.0
cupy-cuda12x>=13.0.0
rapids-cudf>=24.4.0
dask[complete]>=2024.4.0
ray[default]>=2.10.0

# Visualization - Maximum
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.20.0
bokeh>=3.4.0
altair>=5.3.0
holoviews>=1.18.0
panel>=1.4.0
streamlit>=1.33.0
gradio>=4.28.0

# Database - Maximum
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pymongo>=4.7.0
redis>=5.0.0
neo4j>=5.19.0
aioredis>=2.0.1
motor>=3.4.0
asyncpg>=0.29.0
aiomysql>=0.2.0

# API Frameworks - Maximum
fastapi>=0.111.0
strawberry-graphql>=0.225.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
hypercorn>=0.16.0
granian>=1.2.0
litestar>=2.8.0

# Async & Concurrency
asyncio>=3.4.3
aiohttp>=3.9.0
aiofiles>=23.2.0
httpx>=0.27.0
anyio>=4.3.0
trio>=0.25.0
uvloop>=0.19.0

# Monitoring & Observability
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation>=0.45b0
grafana-api>=1.0.3
datadog>=0.49.0
sentry-sdk>=2.0.0

# Security - Maximum
cryptography>=42.0.0
pycryptodome>=3.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.4
pyotp>=2.9.0
python-gnupg>=0.5.2
hashlib>=3.0.0

# Quantum ML Libraries
tensorflow-quantum>=0.7.3
qml>=0.4.0
pennylane-qiskit>=0.35.0
pennylane-cirq>=0.35.0
quantum-tensorflow>=0.1.0

# Neuromorphic Computing
nengo>=3.2.0
brian2>=2.6.0
nest-simulator>=3.6
spinnaker>=1.0.0

# Blockchain & Crypto
web3>=6.17.0
eth-account>=0.11.0
py-solc-x>=2.0.0
bitcoin>=1.1.42

# Document Processing - Maximum
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
lxml>=5.2.0
markdown>=3.6.0
pytesseract>=0.3.10
pdf2image>=1.17.0
camelot-py>=0.11.0

# NLP - Maximum
spacy>=3.7.0
nltk>=3.8.0
gensim>=4.3.0
textblob>=0.18.0
langdetect>=1.0.9
polyglot>=16.7.4
stanza>=1.8.0

# Computer Vision
detectron2>=0.6
mmcv>=2.2.0
mmsegmentation>=1.2.0
mmdetection>=3.3.0
yolov5>=7.0.0
ultralytics>=8.2.0

# Robotics & Control
ros>=2.0.0
pybullet>=3.2.6
gym>=0.26.0
stable-baselines3>=2.3.0
dm-control>=1.0.0

# Time Series
prophet>=1.1.5
statsmodels>=0.14.0
pmdarima>=2.0.0
tslearn>=0.6.0
sktime>=0.28.0

# Graph Processing
networkx>=3.3.0
igraph>=0.11.0
graph-tool>=2.68
networkit>=11.0
snap-stanford>=7.0.0

# Optimization
scipy>=1.13.0
cvxpy>=1.4.0
ortools>=9.9.0
pyomo>=6.7.0
gekko>=1.1.0

# Parallel Processing
joblib>=1.4.0
multiprocess>=0.70.0
pathos>=0.3.0
mpi4py>=3.1.0

# Testing - Maximum
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-xdist>=3.6.0
pytest-benchmark>=4.0.0
hypothesis>=6.100.0
locust>=2.24.0

# Development Tools
black>=24.4.0
ruff>=0.4.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.10.0
pylint>=3.1.0
pre-commit>=3.7.0

# Documentation
sphinx>=7.3.0
mkdocs>=1.6.0
mkdocs-material>=9.5.0
pdoc>=14.4.0

# Profiling & Debugging
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.0
scalene>=1.5.0
viztracer>=0.16.0

# Cloud SDKs
boto3>=1.34.0
google-cloud-storage>=2.16.0
azure-storage-blob>=12.19.0
oci>=2.125.0

# Container & Orchestration
docker>=7.0.0
kubernetes>=29.0.0
helm>=3.14.0

# Additional Advanced Libraries
einops>=0.8.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.16.0
mlflow>=2.12.0
optuna>=3.6.0
hyperopt>=0.2.7
sacred>=0.8.5

# Quantum Cryptography
qcrypto>=1.0.0
quantum-key-distribution>=0.1.0

# Bioinformatics (for DNA storage)
biopython>=1.83
pysam>=0.22.0

# Advanced Mathematics
sympy>=1.12.0
sage>=10.3.0
gmpy2>=2.2.0

# Hardware Acceleration
cuml>=24.4.0
cugraph>=24.4.0
cucim>=24.4.0
pynvml>=11.5.0

# Edge AI
tflite-runtime>=2.14.0
onnxruntime-gpu>=1.17.0
tensorrt>=10.0.0
openvino>=2024.1.0

# Quantum Chemistry (for molecular simulation)
pyscf>=2.5.0
qiskit-nature>=0.7.0

# Advanced Compression
zstandard>=0.22.0
lz4>=4.3.0
brotli>=1.1.0

# Real-time Systems
aiortc>=1.8.0
pyrtmidi>=1.5.0

# AR/VR Support
pyopenxr>=1.0.0
openvr>=2.5.0

# Brain-Computer Interface
mne>=1.7.0
pyeeg>=0.6.0
brainflow>=5.12.0